requirements.txt
Flask
Flask-Cors
PyMuPDF
google-generativeai